Update: Add .route extension to the end of all route files
