+++ src/routes/cookies.routes.js
