--- src/routes/footer/index.js
@@ -1,5 +1,0 @@
-import { contactUs } from './contact-us.js'
-
-export const footerRoutes = [
-  contactUs
-]
