+++ src/routes/static-assets.routes.js
